requirements.txt
apprise
Flask
python-dotenv
selenium
webdriver_manager